requirements.txt
numpy==2.0.*
torch==2.3.*
matplotlib==3.9.*
hydra_core==1.3.*